Fix selection not working
There are two issues. (1) The fix for duplicate commit
messages (https://github.com/ghosty141/consult-git-log-grep/commit/2dbedc0efc90c8a0311c9f2244596fe1ee3eddd4)
changed the result to return a list, but consult--lookup-cdr requires an association list, so the fix is to
put the result inside a list.

(2) The result returned by consult--read is no longer a list, so the car is not needed.